fixed (hopefully) TERM issues
tmux must have a TERM whose name start with `tmux`. This must be
set in tmux.conf with the proper option.

The TERM outside tmux must be the rightone for the terminal used:
for xterm it should be an xterm terminfo, for urxvt must be one
from the rxvt-family.

The actual configuration:
tmux -> tmux-256color-italic (derived from screen!)
outside -> rxvt-unicode-256color